Rebase and cleanup on main (#224)
---------
