+++ .github/workflows/on_pull_request.yaml
@@ -1,5 +1,9 @@
 name: On Pull Request
 
+# On pull_request, we:
+# * run checks
+# * create backport labels if it is against main, only when the PR is opened/reopened
+
 on:
   pull_request:
 
@@ -9,3 +13,12 @@ jobs:
     name: Run Checks
     uses: ./.github/workflows/integrate.yaml
     secrets: inherit
+
+  populate-labels:
+    name: Populate labels
+    if: github.base_ref == 'main' && (github.event.action == 'opened' || github.event.action == 'reopened')
+    uses: canonical/charmed-kubeflow-workflows/.github/workflows/populate-labels.yaml@main
+    secrets: inherit
+    with:
+      track_file_path: ".github/automatic_backport_tracks.yaml"
+      label_prefix: "backport "
