--- modules/kubeflow-ambient/tox.ini
@@ -1,37 +1,0 @@
-# Copyright 2025 Canonical Ltd.
-# See LICENSE file for licensing details.
-
-[tox]
-skipsdist = True
-envlist = tflint, test_deployment
-
-[vars]
-tst_path = {toxinidir}/tests
-
-[testenv]
-passenv =
-    KUBECONFIG
-    PYTHONPATH
-setenv =
-    # Needed for juju cli to work correctly
-    HOME={env:HOME}
-
-[testenv:tflint]
-description = Check Terraform code against coding style standards
-allowlist_externals =
-    tflint
-commands =
-    tflint --recursive
-
-[testenv:test_deployment]
-commands =
-    pytest -v --tb native --asyncio-mode=auto tests/test_deployment.py --log-cli-level=INFO -s {posargs}
-deps =
-    aiohttp
-    lightkube
-    tenacity
-    jubilant
-    juju<4.0.0
-    pytest-dependency
-    pytest-asyncio
-description = Test bundle deployment
