--- modules/kubeflow-mlflow/integrations.tf
@@ -1,102 +1,0 @@
-resource "juju_integration" "mlflow_server_resource_dispatcher_secrets" {
-  model = module.kubeflow.model
-
-  application {
-    name     = module.mlflow.mlflow_server.app_name
-    endpoint = module.mlflow.mlflow_server.requires.secrets
-  }
-
-  application {
-    name     = module.resource_dispatcher.app_name
-    endpoint = module.resource_dispatcher.provides.secrets
-  }
-}
-
-resource "juju_integration" "mlflow_server_resource_dispatcher_pod_defaults" {
-  model = module.kubeflow.model
-
-  application {
-    name     = module.mlflow.mlflow_server.app_name
-    endpoint = module.mlflow.mlflow_server.requires.pod_defaults
-  }
-
-  application {
-    name     = module.resource_dispatcher.app_name
-    endpoint = module.resource_dispatcher.provides.pod_defaults
-  }
-}
-
-resource "juju_integration" "mlflow_server_dashboard_links_provider_links" {
-  count = var.mlflow_dashboard_link ? 1 : 0
-  model = module.kubeflow.model
-
-  application {
-    name     = module.mlflow.mlflow_server.app_name
-    endpoint = module.mlflow.mlflow_server.requires.dashboard_links
-  }
-
-  application {
-    name     = module.kubeflow.dashboard_links_provider.app_name
-    endpoint = module.kubeflow.dashboard_links_provider.provides.links
-  }
-}
-
-resource "juju_integration" "mlflow_server_ingress_provider_ingress" {
-  count = var.mlflow_dashboard_link ? 1 : 0
-  model = module.kubeflow.model
-
-  application {
-    name     = module.mlflow.mlflow_server.app_name
-    endpoint = module.mlflow.mlflow_server.requires.ingress
-  }
-
-  application {
-    name     = module.kubeflow.ingress_provider.app_name
-    endpoint = module.kubeflow.ingress_provider.provides.ingress
-  }
-}
-
-resource "juju_integration" "mlflow_minio_kserve_controller_object_storage" {
-  count = var.mlflow_kserve_integration ? 1 : 0
-  model = module.kubeflow.model
-
-  application {
-    name     = module.mlflow.mlflow_minio.app_name
-    endpoint = module.mlflow.mlflow_minio.provides.object_storage
-  }
-
-  application {
-    name     = module.kubeflow.kserve_controller.app_name
-    endpoint = module.kubeflow.kserve_controller.requires.object_storage
-  }
-}
-
-resource "juju_integration" "kserve_controller_resource_dispatcher_service_accounts" {
-  count = var.mlflow_kserve_integration ? 1 : 0
-  model = module.kubeflow.model
-
-  application {
-    name     = module.kubeflow.kserve_controller.app_name
-    endpoint = module.kubeflow.kserve_controller.requires.service_accounts
-  }
-
-  application {
-    name     = module.resource_dispatcher.app_name
-    endpoint = module.resource_dispatcher.provides.service_accounts
-  }
-}
-
-resource "juju_integration" "kserve_controller_resource_dispatcher_secrets" {
-  count = var.mlflow_kserve_integration ? 1 : 0
-  model = module.kubeflow.model
-
-  application {
-    name     = module.kubeflow.kserve_controller.app_name
-    endpoint = module.kubeflow.kserve_controller.requires.secrets
-  }
-
-  application {
-    name     = module.resource_dispatcher.app_name
-    endpoint = module.resource_dispatcher.provides.secrets
-  }
-}
