--- modules/kubeflow/main.tf
@@ -1,8 +1,0 @@
-resource "juju_model" "kubeflow" {
-  count = var.create_model ? 1 : 0
-  name  = local.model
-}
-
-locals {
-  model = "kubeflow"
-}
