--- modules/kubeflow/tox.ini
@@ -1,12 +1,0 @@
-# Copyright 2024 Canonical Ltd.
-# See LICENSE file for licensing details.
-
-[tox]
-envlist = tflint
-
-[testenv:tflint]
-description = Check Terraform code against coding style standards
-allowlist_externals =
-    tflint
-commands =
-    tflint --recursive
