--- modules/kubeflow/versions.tf
@@ -1,9 +1,0 @@
-terraform {
-  required_version = ">= 1.6"
-  required_providers {
-    juju = {
-      source  = "juju/juju"
-      version = ">= 0.14.0"
-    }
-  }
-}
